# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ideate API
  description: API for the Ideate democratic idea prioritization platform
  version: 0.5.0
  contact:
    name: Ideate Team
  license:
    name: ISC
servers:
  - url: https://idea.surmont.co
    description: Staging
  - url: http://localhost:3000
    description: Development
...
paths:
...
  /api/search:
    get:
      tags: [Search]
      summary: Full-text search
      description: Search across projects and proposals using SQLite FTS5
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search query (minimum 2 characters)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 50
          description: Maximum results to return
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: JWT session cookie (HTTP-only, secure)
  schemas:
...
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        query:
          type: string
      required: [results, query]
    SearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [project, proposal]
        snippet:
          type: string
      required: [id, title, type, snippet]
...